+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 * Set minimum ``click`` version (thank you bildzeitung! `#81 <https://github.com/econchick/interrogate/issues/81>`_, `#82 <https://github.com/econchick/interrogate/pull/82>`_)
+* Add ``--omit-covered-files`` flag to skip reporting fully-covered files (`#85 <https://github.com/econchick/interrogate/issues/85>`_)
 
 .. short-log